+++ assets/components/card-columns.component.jsx
@@ -1,0 +1,47 @@
+import React from 'react';
+import customPropTypes from '../utilities/component/custom-prop-types';
+import getPassThroughProperties from '../utilities/component/get-pass-through-properties';
+import pureRenderShouldComponentUpdate from '../utilities/pure-render-should-component-update';
+
+class CardColumns extends React.Component {
+  constructor(props) {
+    super(props);
+  }
+
+  shouldComponentUpdate(nextProps, nextState) {
+    return pureRenderShouldComponentUpdate(this.props, nextProps, this.state, nextState);
+  }
+
+  getCssClasses() {
+    let cssClasses = ['card-columns'];
+
+    if (this.props.className) {
+      cssClasses = cssClasses.concat(this.props.className.split(' '));
+    }
+
+    return cssClasses;
+  }
+
+  render() {
+    return (
+      <div
+        className={this.getCssClasses().join(' ')}
+        {...getPassThroughProperties(this.props, 'className')}
+      >
+        {this.props.children}
+      </div>
+    );
+  }
+}
+
+CardColumns.displayName = 'CardColumns';
+
+CardColumns.propTypes = {
+  className: React.PropTypes.string
+};
+
+CardColumns.defaultProps = {
+  className: null
+};
+
+export default CardColumns;