+++ assets/components/card-subtitle.component.jsx
@@ -1,0 +1,47 @@
+import React from 'react';
+import customPropTypes from '../utilities/component/custom-prop-types';
+import getPassThroughProperties from '../utilities/component/get-pass-through-properties';
+import pureRenderShouldComponentUpdate from '../utilities/pure-render-should-component-update';
+
+class CardSubtitle extends React.Component {
+  constructor(props) {
+    super(props);
+  }
+
+  shouldComponentUpdate(nextProps, nextState) {
+    return pureRenderShouldComponentUpdate(this.props, nextProps, this.state, nextState);
+  }
+
+  getCssClasses() {
+    let cssClasses = ['card__subtitle'];
+
+    if (this.props.className) {
+      cssClasses = cssClasses.concat(this.props.className.split(' '));
+    }
+
+    return cssClasses;
+  }
+
+  render() {
+    return (
+      <h5
+        className={this.getCssClasses().join(' ')}
+        {...getPassThroughProperties(this.props, 'className')}
+      >
+        {this.props.children}
+      </h5>
+    );
+  }
+}
+
+CardSubtitle.displayName = 'CardSubtitle';
+
+CardSubtitle.propTypes = {
+  className: React.PropTypes.string
+};
+
+CardSubtitle.defaultProps = {
+  className: null,
+};
+
+export default CardSubtitle;