+++ assets/utilities/component/custom-prop-types.js
@@ -1,3 +1,13 @@
 import React from 'react';
 
-export default {};
+const buttonStyleTypes = React.PropTypes.oneOf(['success', 'info', 'warning', 'danger', 'link']);
+const buttonAligns = React.PropTypes.oneOf(['left', 'right']);
+const gridRowAlignments = React.PropTypes.oneOf(['left', 'center', 'right', 'spaced', 'justify']);
+const cardStyleTypes = React.PropTypes.oneOf(['success', 'info', 'warning', 'danger']);
+
+export default {
+  buttonStyleTypes,
+  buttonAligns,
+  gridRowAlignments,
+  cardStyleTypes
+};