+++ web/app/pages/style-guide/assets/examples/buttons/styles.jsx
@@ -1,0 +1,26 @@
+import React from 'react';
+
+import Button from '../../../../../../../assets/components/button.component.jsx';
+
+class SvgIconsIndicatorsExample extends React.Component {
+  constructor(props) {
+    super(props);
+  }
+
+  render() {
+      return (
+        <span>
+          <Button>Default</Button>
+          <Button styleType="success">Success</Button>
+          <Button styleType="info">Info</Button>
+          <Button styleType="warning">Warning</Button>
+          <Button styleType="danger">Danger</Button>
+          <Button styleType="link">Link</Button>
+        </span>
+      );
+  }
+}
+
+SvgIconsIndicatorsExample.displayName = 'SvgIconsIndicatorsExample';
+
+export default SvgIconsIndicatorsExample;