+++ web/app/pages/style-guide/assets/examples/svg-icons/default.jsx
@@ -3,15 +3,15 @@ import React from 'react';
 import SvgIcon from '../../../../../../../assets/components/svg-icon.component.jsx';
 
 class SvgIconsIndicatorsExample extends React.Component {
-    constructor(props) {
-        super(props);
-    }
+  constructor(props) {
+    super(props);
+  }
 
-    render() {
-        return (
-            <SvgIcon fragment="user" />
-        );
-    }
+  render() {
+    return (
+      <SvgIcon fragment="user" />
+    );
+  }
 }
 
 SvgIconsIndicatorsExample.displayName = 'SvgIconsIndicatorsExample';