+++ web/app/pages/style-guide/assets/examples/svg-icons/styles.jsx
@@ -3,32 +3,32 @@ import React from 'react';
 import SvgIcon from '../../../../../../../assets/components/svg-icon.component.jsx';
 
 class SvgIconsIndicatorsExample extends React.Component {
-    constructor(props) {
-        super(props);
-    }
+  constructor(props) {
+    super(props);
+  }
 
-    render() {
-        return (
-            <span>
-                <SvgIcon
-                    styleType="success"
-                    fragment="user"
-                />
-                <SvgIcon
-                    styleType="info"
-                    fragment="user"
-                />
-                <SvgIcon
-                    styleType="warning"
-                    fragment="user"
-                />
-                <SvgIcon
-                    styleType="danger"
-                    fragment="user"
-                />
-            </span>
-        );
-    }
+  render() {
+    return (
+      <span>
+        <SvgIcon
+          styleType="success"
+          fragment="user"
+        />
+        <SvgIcon
+          styleType="info"
+          fragment="user"
+        />
+        <SvgIcon
+          styleType="warning"
+          fragment="user"
+        />
+        <SvgIcon
+          styleType="danger"
+          fragment="user"
+        />
+      </span>
+    );
+  }
 }
 
 SvgIconsIndicatorsExample.displayName = 'SvgIconsIndicatorsExample';