+++ web/app/pages/style-guide/colors.page.jsx
@@ -1,0 +1,33 @@
+import React from 'react';
+
+import CodeExample from '../../react/components/code-example.component.jsx';
+
+// import StylesExample from './assets/examples/buttons/styles.jsx';
+
+// import { readFileSync } from 'fs';
+// import { join } from 'path';
+
+// const stylesExampleContent = readFileSync(join(__dirname, '/assets/examples/buttons/styles.jsx'), 'utf8');
+
+class ButtonsPage extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {};
+  }
+
+  render() {
+    return (
+      <div className="p-style-guide-colors">
+        <h1>Color</h1>
+      </div>
+    );
+  }
+}
+
+ButtonsPage.displayName = 'ButtonsPage';
+
+ButtonsPage.contextTypes = {
+  router: React.PropTypes.object.isRequired
+};
+
+export default ButtonsPage;